server/requirements.txt
Django==2.2.1
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
psycopg2